+++ CHANGES.md
@@ -1,6 +1,12 @@
 Optique changelog
 =================
 
+Version 0.7.14
+--------------
+
+To be released.
+
+
 Version 0.7.13
 --------------